fix: correctly strip source file prefixes when no webpack.context is defined
This now uses the same default logic as webpack for identifying the source root.
Based on https://webpack.js.org/configuration/entry-context/#context, if the context is not specified, the current
directory that the process is running from should be assumed instead.

Align util to mimic this behaviour to ensure paths are correctly de-absoluted during processing.